+++ spec/www/spec/db-sql-operations-test.js
@@ -1054,10 +1054,7 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWindows || (isAndroid && isImpl2))
-              expect(error.code).toBe(0);
-            else
-              expect(error.code).toBe(5);
+            expect(error.code).toBe(0);
 
             if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
@@ -1082,10 +1079,7 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWindows || (isAndroid && isImpl2))
-              expect(error.code).toBe(0);
-            else
-              expect(error.code).toBe(5);
+            expect(error.code).toBe(0);
 
             if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
@@ -1110,10 +1104,7 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWindows || (isAndroid && isImpl2))
-              expect(error.code).toBe(0);
-            else
-              expect(error.code).toBe(5);
+            expect(error.code).toBe(0);
 
             if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
@@ -1138,10 +1129,7 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWindows || (isAndroid && isImpl2))
-              expect(error.code).toBe(0);
-            else
-              expect(error.code).toBe(5);
+            expect(error.code).toBe(0);
 
             if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
@@ -1166,10 +1154,7 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWindows || (isAndroid && isImpl2))
-              expect(error.code).toBe(0);
-            else
-              expect(error.code).toBe(5);
+            expect(error.code).toBe(0);
 
             if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
@@ -1194,10 +1179,7 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWindows || (isAndroid && isImpl2))
-              expect(error.code).toBe(0);
-            else
-              expect(error.code).toBe(5);
+            expect(error.code).toBe(0);
 
             if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
@@ -1222,10 +1204,7 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWindows || (isAndroid && isImpl2))
-              expect(error.code).toBe(0);
-            else
-              expect(error.code).toBe(5);
+            expect(error.code).toBe(0);
 
             if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
@@ -1469,10 +1448,7 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWindows || (isAndroid && isImpl2))
-              expect(error.code).toBe(0);
-            else
-              expect(error.code).toBe(5);
+            expect(error.code).toBe(0);
 
             if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
@@ -1493,10 +1469,7 @@ var mytests = function() {
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
 
-            if (isWindows || (isAndroid && isImpl2))
-              expect(error.code).toBe(0);
-            else
-              expect(error.code).toBe(5);
+            expect(error.code).toBe(0);
 
             if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);